get format writing working in a shell script for batch suitesparse matrices
